fix #159; fix #207
